Move audio_utils into core.
